Move building validity tests to separate file
